Use ruby for ansible auth decrypt
This adds decryption for ansible credentials in ruby, without the need
for external python libs to be installed and an external python
standalone_decrypt.py file to required.

When developing this, I was specifically comparing the values being
generated for what I was writing up against what was being generated by
the standalone_decrypt.py code (and underlying libs), so I am pretty
confident I am at least 90% of the way there with this decrypt code
here.